Generalize node init for use with full node
